+++ languagetool-language-modules/fr/src/main/java/org/languagetool/language/French.java
@@ -19,11 +19,13 @@
 package org.languagetool.language;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.ResourceBundle;
 import java.io.IOException;
 
+import org.languagetool.GlobalConfig;
 import org.languagetool.Language;
 import org.languagetool.LanguageMaintainedState;
 import org.languagetool.UserConfig;
@@ -104,7 +106,7 @@ public class French extends Language implements AutoCloseable {
   }
 
   @Override
-  public List<Rule> getRelevantRules(ResourceBundle messages, UserConfig userConfig, List<Language> altLanguages) throws IOException {
+  public List<Rule> getRelevantRules(ResourceBundle messages, UserConfig userConfig, Language motherTongue, List<Language> altLanguages) throws IOException {
     return Arrays.asList(
             new CommaWhitespaceRule(messages),
             new DoublePunctuationRule(messages),
@@ -122,8 +124,18 @@ public class French extends Language implements AutoCloseable {
             new SentenceWhitespaceRule(messages),
             // specific to French:
             new CompoundRule(messages),
+            new QuestionWhitespaceStrictRule(messages, this),
             new QuestionWhitespaceRule(messages, this)
     );
+  }
+
+  @Override
+  public List<Rule> getRelevantRulesGlobalConfig(ResourceBundle messages, GlobalConfig globalConfig, UserConfig userConfig, Language motherTongue, List<Language> altLanguages) throws IOException {
+    List<Rule> rules = new ArrayList<>();
+    if (globalConfig != null && globalConfig.getGrammalecteServer() != null) {
+      rules.add(new GrammalecteRule(messages, globalConfig));
+    }
+    return rules;
   }
 
   /** @since 3.1 */
@@ -138,7 +150,7 @@ public class French extends Language implements AutoCloseable {
   /** @since 3.1 */
   @Override
   public List<Rule> getRelevantLanguageModelRules(ResourceBundle messages, LanguageModel languageModel) throws IOException {
-    return Arrays.<Rule>asList(
+    return Arrays.asList(
             new FrenchConfusionProbabilityRule(messages, languageModel, this)
     );
   }
@@ -157,6 +169,18 @@ public class French extends Language implements AutoCloseable {
   @Override
   public LanguageMaintainedState getMaintainedState() {
     return LanguageMaintainedState.ActivelyMaintained;
+  }
+
+  @Override
+  public int getPriorityForId(String id) {
+    switch (id) {
+      case "FRENCH_WHITESPACE_STRICT": return 1;  // default off, but if on, it should overwrite FRENCH_WHITESPACE 
+      case "FRENCH_WHITESPACE": return 0;
+    }
+    if (id.startsWith("grammalecte_")) {
+      return -1;
+    }
+    return super.getPriorityForId(id);
   }
 
 }