+++ languagetool-server/src/main/java/org/languagetool/server/ApiV2.java
@@ -38,7 +38,7 @@ import java.io.StringWriter;
 import java.net.HttpURLConnection;
 import java.util.*;
 
-import static org.languagetool.server.ServerTools.print;
+import static org.languagetool.server.LanguageToolHttpHandler.API_DOC_URL;
 
 /**
  * Handle requests to {@code /v2/} of the HTTP API. 
@@ -76,7 +76,7 @@ class ApiV2 {
       // private (i.e. undocumented) API for our own use only
       handleLogRequest(httpExchange, parameters);
     } else {
-      throw new PathNotFoundException("Unsupported action: '" + path + "'");
+      throw new PathNotFoundException("Unsupported action: '" + path + "'. Please see " + API_DOC_URL);
     }
   }
 
@@ -85,6 +85,7 @@ class ApiV2 {
     ServerTools.setCommonHeaders(httpExchange, JSON_CONTENT_TYPE, allowOriginUrl);
     httpExchange.sendResponseHeaders(HttpURLConnection.HTTP_OK, response.getBytes(ENCODING).length);
     httpExchange.getResponseBody().write(response.getBytes(ENCODING));
+    ServerMetricsCollector.getInstance().logResponse(HttpURLConnection.HTTP_OK);
   }
 
   private void handleCheckRequest(HttpExchange httpExchange, Map<String, String> parameters, ErrorRequestLimiter errorRequestLimiter, String remoteAddress) throws Exception {
@@ -245,6 +246,7 @@ class ApiV2 {
     ServerTools.setCommonHeaders(httpExchange, JSON_CONTENT_TYPE, allowOriginUrl);
     httpExchange.sendResponseHeaders(HttpURLConnection.HTTP_OK, response.getBytes(ENCODING).length);
     httpExchange.getResponseBody().write(response.getBytes(ENCODING));
+    ServerMetricsCollector.getInstance().logResponse(HttpURLConnection.HTTP_OK);
   }
 
   private void handleLogRequest(HttpExchange httpExchange, Map<String, String> parameters) throws IOException {
@@ -257,6 +259,7 @@ class ApiV2 {
     String response = "OK";
     httpExchange.sendResponseHeaders(HttpURLConnection.HTTP_OK, response.getBytes(ENCODING).length);
     httpExchange.getResponseBody().write(response.getBytes(ENCODING));
+    ServerMetricsCollector.getInstance().logResponse(HttpURLConnection.HTTP_OK);
   }
 
   private AnnotatedText getAnnotatedTextFromString(JsonNode data, String text) {