+++ languagetool-server/src/main/java/org/languagetool/server/TextChecker.java
@@ -52,7 +52,6 @@ import static org.languagetool.server.ServerTools.print;
  */
 abstract class TextChecker {
 
-
   protected abstract void setHeaders(HttpExchange httpExchange);
   protected abstract String getResponse(AnnotatedText text, DetectedLanguage lang, Language motherTongue, List<RuleMatch> matches,
                                         List<RuleMatch> hiddenMatches, String incompleteResultReason);
@@ -102,6 +101,13 @@ abstract class TextChecker {
       this.logServerId = null;
     }
 
+    ServerMetricsCollector.getInstance().logHiddenServerConfiguration(config.getHiddenMatchesServer() != null);
+
+    if (cache != null) {
+      ServerMetricsCollector.getInstance().monitorCache("languagetool_matches_cache", cache.getMatchesCache());
+      ServerMetricsCollector.getInstance().monitorCache("languagetool_sentences_cache", cache.getSentenceCache());
+    }
+
     pipelinePool = new PipelinePool(config, cache, internalServer);
     if (config.isPipelinePrewarmingEnabled()) {
       ServerTools.print("Prewarming pipelines...");
@@ -137,12 +143,12 @@ abstract class TextChecker {
         QueryParams params = new QueryParams(Collections.emptyList(), Collections.emptyList(), addonDisabledRules,
           Collections.emptyList(), Collections.emptyList(), false, true,
           false, false, mode, null);
-        PipelinePool.PipelineSettings settings = new PipelinePool.PipelineSettings(language, null, params, user);
+        PipelinePool.PipelineSettings settings = new PipelinePool.PipelineSettings(language, null, params, config.globalConfig, user);
         prewarmSettings.put(settings, NUM_PIPELINES_PER_SETTING);
 
-        PipelinePool.PipelineSettings settingsMotherTongueEqual = new PipelinePool.PipelineSettings(language, language, params, user);
+        PipelinePool.PipelineSettings settingsMotherTongueEqual = new PipelinePool.PipelineSettings(language, language, params, config.globalConfig, user);
         PipelinePool.PipelineSettings settingsMotherTongueEnglish = new PipelinePool.PipelineSettings(language,
-          Languages.getLanguageForName("English"), params, user);
+          Languages.getLanguageForName("English"), params, config.globalConfig, user);
         prewarmSettings.put(settingsMotherTongueEqual, NUM_PIPELINES_PER_SETTING);
         prewarmSettings.put(settingsMotherTongueEnglish, NUM_PIPELINES_PER_SETTING);
       }
@@ -192,6 +198,7 @@ abstract class TextChecker {
     if (aText.getPlainText().length() > limits.getMaxTextLength()) {
       String msg = "limit: " + limits.getMaxTextLength() + ", size: " + aText.getPlainText().length();
       logger.log(new DatabaseAccessLimitLogEntry("MaxCharacterSizeExceeded", logServerId, agentId, userId, msg, referrer, userAgent));
+      ServerMetricsCollector.getInstance().logRequestError(ServerMetricsCollector.RequestErrorType.MAX_TEXT_SIZE);
       throw new TextTooLongException("Your text exceeds the limit of " + limits.getMaxTextLength() +
               " characters (it's " + aText.getPlainText().length() + " characters). Please submit a shorter text.");
     }
@@ -208,6 +215,7 @@ abstract class TextChecker {
     boolean autoDetectLanguage = getLanguageAutoDetect(parameters);
     List<String> preferredVariants = getPreferredVariants(parameters);
     if (parameters.get("noopLanguages") != null && !autoDetectLanguage) {
+      ServerMetricsCollector.getInstance().logRequestError(ServerMetricsCollector.RequestErrorType.INVALID_REQUEST);
       throw new IllegalArgumentException("You can specify 'noopLanguages' only when also using 'language=auto'");
     }
     List<String> noopLangs = parameters.get("noopLanguages") != null ?
@@ -233,6 +241,7 @@ abstract class TextChecker {
         Language altLang = Languages.getLanguageForShortCode(langCode);
         altLanguages.add(altLang);
         if (altLang.hasVariant() && !altLang.isVariant()) {
+          ServerMetricsCollector.getInstance().logRequestError(ServerMetricsCollector.RequestErrorType.INVALID_REQUEST);
           throw new IllegalArgumentException("You specified altLanguage '" + langCode + "', but for this language you need to specify a variant, e.g. 'en-GB' instead of just 'en'");
         }
       }
@@ -244,9 +253,11 @@ abstract class TextChecker {
     List<CategoryId> disabledCategories = getCategoryIds("disabledCategories", parameters);
 
     if ((disabledRules.size() > 0 || disabledCategories.size() > 0) && useEnabledOnly) {
+      ServerMetricsCollector.getInstance().logRequestError(ServerMetricsCollector.RequestErrorType.INVALID_REQUEST);
       throw new IllegalArgumentException("You cannot specify disabled rules or categories using enabledOnly=true");
     }
     if (enabledRules.isEmpty() && enabledCategories.isEmpty() && useEnabledOnly) {
+      ServerMetricsCollector.getInstance().logRequestError(ServerMetricsCollector.RequestErrorType.INVALID_REQUEST);
       throw new IllegalArgumentException("You must specify enabled rules or categories when using enabledOnly=true");
     }
 
@@ -312,6 +323,7 @@ abstract class TextChecker {
     } catch (ExecutionException e) {
       future.cancel(true);
       if (ExceptionUtils.getRootCause(e) instanceof ErrorRateTooHighException) {
+        ServerMetricsCollector.getInstance().logRequestError(ServerMetricsCollector.RequestErrorType.TOO_MANY_ERRORS);
         logger.log(new DatabaseCheckErrorLogEntry("ErrorRateTooHigh", logServerId, agentId, userId, lang, detLang.getDetectedLanguage(), textSize, "matches: " + ruleMatchesSoFar.size()));
       }
       if (params.allowIncompleteResults && ExceptionUtils.getRootCause(e) instanceof ErrorRateTooHighException) {
@@ -344,6 +356,7 @@ abstract class TextChecker {
         incompleteResultReason = "Results are incomplete: text checking took longer than allowed maximum of " + 
                 String.format(Locale.ENGLISH, "%.2f", limits.getMaxCheckTimeMillis()/1000.0) + " seconds";
       } else {
+        ServerMetricsCollector.getInstance().logRequestError(ServerMetricsCollector.RequestErrorType.MAX_CHECK_TIME);
         logger.log(new DatabaseCheckErrorLogEntry("MaxCheckTimeExceeded",
           logServerId, agentId, limits.getPremiumUid(), lang, detLang.getDetectedLanguage(), textSize, "load: "+ loadInfo));
         throw new RuntimeException(message, e);
@@ -357,6 +370,7 @@ abstract class TextChecker {
       config.getHiddenMatchesLanguages().contains(lang)) {
       if(config.getHiddenMatchesServerFailTimeout() > 0 && lastHiddenMatchesServerTimeout != -1 &&
         System.currentTimeMillis() - lastHiddenMatchesServerTimeout < config.getHiddenMatchesServerFailTimeout()) {
+        ServerMetricsCollector.getInstance().logHiddenServerStatus(false);
         print("Warn: Skipped querying hidden matches server at " +
           config.getHiddenMatchesServer() + " because of recent error/timeout (timeout=" + config.getHiddenMatchesServerFailTimeout() + "ms).");
       } else {
@@ -366,9 +380,11 @@ abstract class TextChecker {
           List<RemoteRuleMatch> extensionMatches = resultExtender.getExtensionMatches(aText.getPlainText(), parameters);
           hiddenMatches = resultExtender.getFilteredExtensionMatches(matches, extensionMatches);
           long end = System.currentTimeMillis();
-          print("Hidden matches: " + extensionMatches.size() + " -> " + hiddenMatches.size() + " in " + (end - start) + "ms");
+          print("Hidden matches: " + extensionMatches.size() + " -> " + hiddenMatches.size() + " in " + (end - start) + "ms for " + lang.getShortCodeWithCountryAndVariant());
+          ServerMetricsCollector.getInstance().logHiddenServerStatus(true);
           lastHiddenMatchesServerTimeout = -1;
         } catch (Exception e) {
+          ServerMetricsCollector.getInstance().logHiddenServerStatus(false);
           print("Warn: Failed to query hidden matches server at " + config.getHiddenMatchesServer() + ": " + e.getClass() + ": " + e.getMessage());
           lastHiddenMatchesServerTimeout = System.currentTimeMillis();
         }
@@ -385,6 +401,7 @@ abstract class TextChecker {
     try {
       httpExchange.sendResponseHeaders(HttpURLConnection.HTTP_OK, response.getBytes(ENCODING).length);
       httpExchange.getResponseBody().write(response.getBytes(ENCODING));
+      ServerMetricsCollector.getInstance().logResponse(HttpURLConnection.HTTP_OK);
     } catch (IOException exception) {
       // the client is disconnected
       messageSent = "notSent: " + exception.getMessage();
@@ -406,17 +423,22 @@ abstract class TextChecker {
             + ", m:" + mode.toString().toLowerCase());
 
     int matchCount = matches.size();
-    DatabaseCheckLogEntry logEntry = new DatabaseCheckLogEntry(userId, agentId, logServerId, textSize, matchCount,
-      lang, detLang.getDetectedLanguage(), computationTime, textSessionId, mode.toString());
     Map<String, Integer> ruleMatchCount = new HashMap<>();
-    if (!config.isSkipLoggingRuleMatches()) {
-      for (RuleMatch match : matches) {
-        String ruleId = match.getRule().getId();
-        ruleMatchCount.put(ruleId, ruleMatchCount.getOrDefault(ruleId, 0) + 1);
-      }
+    for (RuleMatch match : matches) {
+      String ruleId = match.getRule().getId();
+      ruleMatchCount.put(ruleId, ruleMatchCount.getOrDefault(ruleId, 0) + 1);
     }
-    logEntry.setRuleMatches(new DatabaseRuleMatchLogEntry(ruleMatchCount));
-    logger.log(logEntry);
+
+    ServerMetricsCollector.getInstance().logCheck(
+      lang, computationTime, textSize, matchCount, mode, agent, ruleMatchCount);
+
+    if (!config.isSkipLoggingChecks()) {
+      DatabaseCheckLogEntry logEntry = new DatabaseCheckLogEntry(userId, agentId, logServerId, textSize, matchCount,
+        lang, detLang.getDetectedLanguage(), computationTime, textSessionId, mode.toString());
+      logEntry.setRuleMatches(new DatabaseRuleMatchLogEntry(
+        config.isSkipLoggingRuleMatches() ? Collections.emptyMap() : ruleMatchCount));
+      logger.log(logEntry);
+    }
   }
 
   private List<String> getUserDictWords(Long userId) {
@@ -440,7 +462,7 @@ abstract class TextChecker {
       //print("Matches    : " + cache.getMatchesCache().stats().hitRate() + " hit rate");
       //print("Sentences  : " + cache.getSentenceCache().stats().hitRate() + " hit rate");
       //print("Size       : " + cache.getMatchesCache().size() + " (matches cache), " + cache.getSentenceCache().size() + " (sentence cache)");
-      logger.log(new DatabaseCacheStatsLogEntry(logServerId, (float) hitRate));
+      //logger.log(new DatabaseCacheStatsLogEntry(logServerId, (float) hitRate));
     }
     PipelinePool.PipelineSettings settings = null;
 
@@ -456,7 +478,7 @@ abstract class TextChecker {
     } else {
       Pipeline lt = null;
       try {
-        settings = new PipelinePool.PipelineSettings(lang, motherTongue, params, userConfig);
+        settings = new PipelinePool.PipelineSettings(lang, motherTongue, params, config.globalConfig, userConfig);
         lt = pipelinePool.getPipeline(settings);
         return lt.check(aText, true, JLanguageTool.ParagraphHandling.NORMAL, listener, params.mode);
       } finally {